% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bench_calcAUC.R
\name{bench_calcAUC}
\alias{bench_calcAUC}
\title{Benchmarking: Calculate of the AUC of different cases}
\usage{
bench_calcAUC(out_list, cluster, cell_types)
}
\arguments{
\item{out_list}{The list of prediction probability matrix of each case}

\item{cluster}{The ground-truth}

\item{cell_types}{Character vector of cell types of the data}
}
\value{
A dataset of AUC per cell type of each case
}
\description{
This function calculates the AUC of different cases
}
