% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bench_MarkerDetect.R
\name{bench_MarkerDetect}
\alias{bench_MarkerDetect}
\title{Benchmarking of the marker detection ability}
\usage{
bench_MarkerDetect(
  threshold,
  input_list,
  ref_list,
  xlab = NULL,
  ylab = NULL,
  title = NULL,
  legend_title = NULL,
  round_by = 3
)
}
\arguments{
\item{input_list}{List combining several outputs of ds_feature_importance}

\item{ref_list}{List of reference markers}

\item{xlab}{The text for the x axis name}

\item{ylab}{The text for the y axis name}

\item{title}{The text for the title}

\item{legend_title}{The text fot the legend title}

\item{round_by}{Integer indicating the number of decimal places}
}
\value{
A heatmap with the marker detection rate per case
}
\description{
This function calculates the marker detection rate per case
}
