% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds_correlation_plot.R
\name{ds_correlation_plot}
\alias{ds_correlation_plot}
\title{This function plots the correlation matrix between two datasets}
\usage{
ds_correlation_plot(x.data, x.assay, y.data, y.assay, downsample, features)
}
\arguments{
\item{x.data}{The Seurat Object that contains the first data}

\item{x.assay}{The assay in the Seurat Object that contains the first data}

\item{y.data}{The Seurat Object that contains the second data}

\item{y.assay}{The assay in the Seurat Object that contains the second data}

\item{downsample}{The number of cells to downsample in each class}

\item{markers}{The markers to take into account in the correlation matrix}
}
\description{
This function generates a correlation plot
}
