% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds_plot_importance.R
\name{ds_plot_importance}
\alias{ds_plot_importance}
\title{This function plots the importance of the features}
\usage{
ds_plot_importance(
  data,
  n = 30,
  Decrease = TRUE,
  general_alteration = TRUE,
  main = ""
)
}
\arguments{
\item{data}{The output of \code{ds_feature_importance}.}

\item{n}{The number of features to be plotted (Default=30)}

\item{Decrease}{If the importance are plotted decreasing or increasing.}

\item{general_alteration}{If plot the absolut scores, if \code{TRUE}, the plot would not take into account if the
alteration means an increase or decrease. (Default=TRUE)}

\item{main}{The title of the plot}
}
\value{
A feature importance plot
}
\description{
This function plots the n top importance features decreasing or increasing.
}
\examples{
ds_plot_importance(data,n=50,Decrease=TRUE,general_alteration=FALSE,main="Feature Importance Plot")

}
