% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds_plot_performance.R
\name{ds_plot_performance}
\alias{ds_plot_performance}
\title{Plot performance as radarchart/spider plot}
\usage{
ds_plot_performance(data, parameters, legend = TRUE, colors, plwd = 1)
}
\arguments{
\item{data}{Output of \code{ds_performance()}}

\item{parameters}{The parameters that will be plotted}

\item{legend}{Plot with legend (Default=TRUE)}

\item{colors}{Vectors of colors for the plot}

\item{plwd}{Line width of the plot}
}
\value{
A radarchart plot
}
\description{
This function plots the performance calculated by \code{ds_performance()}
}
\examples{
#not run
ds_plot_performance(data = performance$Performance,
  parameters = c("Sensitivity","Specificity","Accuracy","Classification Rate"),
  colors = colors_border1,legend = FALSE)
}
