% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds_prepare_data.R
\name{ds_prepare_data}
\alias{ds_prepare_data}
\title{This function prepares the needed inputs to train the dnn model}
\usage{
ds_prepare_data(
  ref.data,
  ref.assay,
  query.data,
  query.assay,
  markers,
  markers_data
)
}
\arguments{
\item{ref.data}{The Seurat Object that contains the reference data}

\item{ref.assay}{The assay in the Seurat Object that contains the reference data}

\item{query.data}{The Seurat Object that contains the query data}

\item{query.assay}{The assay in the Seurat Object that contains the query data}

\item{markers}{Pre-defined marker genes. If missing, \code{markers_data} is required, otherwise, the next parameters must be empty.}

\item{markers_data}{The output of the R function \code{FindAllMarkers} provided by \code{"Seurat"}. If not provided, the function will automatically calculate it.}
}
\value{
The list containing all input data needed to train the dnn model.
}
\description{
This function generates a list with the needed inputs to generate the dnn model.
}
\examples{
#not run
no_markers_no_top_markers<-ds_prepare_data(ref.data,ref.assay,query.data,query.assay,markers_data,top_markers=F)
no_markers_top_markers<-ds_prepare_data(ref.data,ref.assay,query.data,query.assay,markers_data,top_markers=T,ntop=100)
markers<-ds_prepare_data(ref.data,ref.assay,query.data,query.assay,markers)

}
