% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds_split_data_dnn_1.R
\name{ds_split_data_dnn_1}
\alias{ds_split_data_dnn_1}
\title{This function split the reference data in training and test set to use it as input for the ds_dnn_model function.}
\usage{
ds_split_data_dnn_1(
  scale.data,
  clus,
  genes,
  prop = NULL,
  verbose = TRUE,
  scaled_data = FALSE,
  sct_norm = TRUE,
  train.samples = NULL
)
}
\arguments{
\item{scale.data}{scale.data A scaled/normalized matrix of gene expressions like in the \code{scale.data}
of the Seurat object. Rows are genes and columns are cells from the reference
dataset.}

\item{clus}{A named factor with the cell type annotations from the reference data.}

\item{genes}{The set of genes you want to use.}

\item{prop}{The proportion of cells for the training dataset (default=0.5).}

\item{verbose}{Logical, controls the displaying of additional messages while
running the function. Defaults to \code{TRUE}.}

\item{scaled_data}{Set according to the normalizing method of the data. (default=TRUE)}

\item{train.samples}{Set train samples}

\item{sct_nrom}{Set according to the normalizing method of the data. (default=TRUE)}
}
\value{
A list with the following data: train_x= training_data,train_y= categorical cell type variables,
test_x =test_data,test_y=categorical celltype from the test data, classes=original cell types, train_samples=Set of train samples).
}
\description{
This function split the reference data in training and test set to use it as input for the ds_dnn_model function.
}
\examples{
#not run
}
